feat(client): add proxy object for `request.agent` for compatibility
In node environments, the code in `node/agent.js` creates a proxy
object around `Agent` so that it can be instantiated without using
the `new` keyword. The equivalent code for the browser version
(in `client.js`) defines `request.agent` as `() => new Agent()`.

In supertest, the `TestAgent` attempts to subclass `Agent` with the
call `Object.setPrototypeOf(TestAgent.prototype, Agent.prototype);`.
This works in the node environment but not in the browser version,
because in the browser version `() => new Agent()` is not an object
and has an undefined prototype, leading to the error message
`Object prototype may only be an Object or null: undefined`.

Wrap `request.agent` in a proxy object for the client/browser case
so that supertest can also use the library in browser contexts
without error.